requirements.txt
streamlit
requests
selenium
webdriver_manager
langchain
langchain_openai
openai
pillow
pytest
python-dotenv
streamlit_chat
fastapi
pydantic
uvicorn
langchain
pytesseract
bs4
langgraph
autogen
langchain-community
google-auth-httplib2
google-api-python-client
langchain_experimental